LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-audio-background-expected.txt

Testing that we show a gray background for macOS audio elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.querySelector('.controls-bar') instanceof HTMLDivElement became true
PASS getComputedStyle(mediaControls.querySelector('.controls-bar')).backgroundColor is "rgb(41, 41, 41)"
PASS getComputedStyle(mediaControls.querySelector('.controls-bar > .background-tint')).display is "none"
PASS successfullyParsed is true

TEST COMPLETE
